Task_3/resume_screening/requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
langsmith
python-dotenv
